got location working
it now can find your current location and relay it to a url

i don't really want to have a manual select for the location so
if the don't agree to let me see their location it wont work.

TODO: i may come back eventually and add that functionality but idk.