put in a no script prompt and some little edits
since the open weather app is not https it will not work on code pen properly
In modern browsers, it is considered unsafe script due to it being a unsecured
HTTP connection so the javascript is considered unsafe.
the only way to get it working is to load unsafe script.